reduce exported API to single method: NewContextWithServerTransportStream (and corresponding interface); move to grpc package (out of transport package)
