+++ server_test.go
@@ -25,7 +25,9 @@ import (
 	"testing"
 	"time"
 
+	"golang.org/x/net/context"
 	"google.golang.org/grpc/test/leakcheck"
+	"google.golang.org/grpc/transport"
 )
 
 type emptyServiceServer interface{}
@@ -120,5 +122,15 @@ func TestGetServiceInfo(t *testing.T) {
 
 	if !reflect.DeepEqual(info, want) {
 		t.Errorf("GetServiceInfo() = %+v, want %+v", info, want)
+	}
+}
+
+func TestStreamContext(t *testing.T) {
+	expectedStream := &transport.Stream{}
+	ctx := NewContextWithServerTransportStream(context.Background(), expectedStream)
+	s := serverTransportStreamFromContext(ctx)
+	stream, ok := s.(*transport.Stream)
+	if !ok || expectedStream != stream {
+		t.Fatalf("GetStreamFromContext(%v) = %v, %t, want: %v, true", ctx, stream, ok, expectedStream)
 	}
 }
